Fix bug in routine where request body is deleted.
Repair a bug in the self._make_auth() routine where the OAuth1 client is generated before a request body is available. This causes all SuiteQL queries to fail, along with any other functions that require a request body. Resolve by storing the connection details in the class and calling that method again when we detect a post body.