Add ability to include additional headers in a search request
This allows me to add headers such as `{'searchPreferences': {'bodyFieldsOnly': False}}` so that I can get a more detailed response back from the Search action.

My use case was getting the list of values back from a CustomList object. The only way to do that is to include those headers.